requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
email-validator>=2.0.0
python-dotenv==1.0.0
pytest==7.4.4
httpx==0.26.0
python-dateutil==2.8.2

# RAG-specific
# Pin numpy to < 2.0 for ChromaDB compatibility
numpy<2.0.0
chromadb==0.4.22
openai==1.10.0
sentence-transformers==2.3.1

